move helper methods to suite file
